requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
croniter==1.0.15
dj-database-url==0.5.0
Django==3.2.5
django-admin==2.0.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-on-heroku==1.1.2
django-six==1.0.4
django-tables2==2.4.0
idna==3.2
psycopg2==2.8
psycopg2-binary==2.8.5
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
xlwt==1.3.0
